requirements.txt
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.6.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
Flask-SocketIO==5.3.4
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==1.0.0
python-engineio==4.4.1
python-socketio==5.8.0
requests==2.31.0
simple-websocket==0.10.0
soupsieve==2.4.1
urllib3==2.0.2
Werkzeug==2.3.4
wsproto==1.2.0
